YAL: undo rename of pushmessages_optin.py
